Minnor fixes so some files
major overhaul of loader program. Imports will now be done automatically.

To edit which files are run edit plugins.txt

I've kept the old way of running for telnet, ssh, and https but that can be removed later.

I'd love to have a talk with you guys about standarizing method name for initiation of program class to server_plugin.

http_reader2.py is basically the same thing as http_reader.py but on port 81 to test new format of files.

Need to work on locks to allow console prints statements to not go one on top of the other.